Set up the log file
