+++ server/db.ts
@@ -1,39 +1,18 @@
 import * as path from 'path'
 import { promises as fs } from 'fs'
 
-import {
-  Generated,
-  Insertable,
-  Selectable,
-  Kysely,
-  Migrator,
-  FileMigrationProvider,
-  ColumnType,
-  sql,
-} from 'kysely'
+import { Kysely, Migrator, FileMigrationProvider, sql } from 'kysely'
 import { BunSqliteDialect } from 'kysely-bun-worker/normal'
 import debug from 'debug'
+import { Schema } from './db-schema'
 
 const d = debug('ha:db')
 
-export type Timestamp = ColumnType<Date, Date | string, Date | string>
-
-export interface Schema {
-  signals: SignalTable
-}
-
-export interface SignalTable {
-  id: Generated<number>
-  createdAt: Generated<Timestamp>
-  automationHash: string
-}
-
-export type Signal = Selectable<SignalTable>
-export type NewSignal = Insertable<SignalTable>
-
 export async function createDatabase() {
-  const dbPath = process.env.DATA_DIR ? path.join(process.env.DATA_DIR, 'app.db') : './app.db'
-  
+  const dbPath = process.env.DATA_DIR
+    ? path.join(process.env.DATA_DIR, 'app.db')
+    : './app.db'
+
   const db = new Kysely<Schema>({
     dialect: new BunSqliteDialect({ url: dbPath }),
     log(ev) {
@@ -57,12 +36,4 @@ export async function createDatabase() {
 
   await migrator.migrateToLatest()
   return db
-}
-
-export async function test() {
-  const db = await createDatabase()
-  await db
-    .insertInto('signals')
-    .values({ automationHash: 'foo' })
-    .executeTakeFirst()
 }
